# doc/api.yaml
openapi: 3.0.0
info:
  title: Home Library Service
  description: Home music library service
  version: 1.0.0
servers:
  - url: /
components:
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: 0a35dd62-e09f-444b-a628-f4e7c6954f57
        login:
          type: string
          example: 'TestUser'
        version:
          type: number
          example: 1
        createAt:
          type: number
          example: 1655000000
        updateAt:
          type: number
          example: 1655000000
      required:
        - id
        - login
        - version
        - createdAt
        - updatedAt
...
    Album:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          example: 'Innuendo'
        year:
          type: integer
          example: 1991
        artistId:
          type: string
          format: uuid
          nullable: true
      required:
        - id
        - name
        - year
...
paths:
  /user:
    get:
      tags:
        - Users
      summary: Get all users
      description: Gets all users
      responses:
        200:
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
    post:
      tags:
        - Users
      summary: Create user
      description: Creates a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              title: example
              properties:
                login:
                  type: string
                  description: The user's login
                password:
                  type: string
                  description: The user's password
              required:
                - login
                - password
      responses:
        201:
          description: The user has been created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        400:
          description: Bad request. body does not contain required fields
          content:
            application/json:
              schema:
                type: object
                title: example
                properties:
                  statusCode:
                    type: number
                    example: 400
                  message:
                    type: string
                    example:
                      - login must be a string
                      - login should not be empty
                      - password must be a string
                      - password should not be empty
                  error:
                    type: string
                    example: Bad Request
  /user/{userId}:
    parameters:
      - name: userId
        in: path
        required: true
        schema:
          type: string
          example: 0a35dd62-e09f-444b-a628-f4e7c6954f57
    get:
      tags:
        - Users
      summary: Get single user by id
      description: Get single user by id
      responses:
        200:
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        400:
          description: Bad request. userId is invalid (not uuid)
          content:
            application/json:
              schema:
                type: object
                title: example
                properties:
                  statusCode:
                    type: number
                    example: 400
                  message:
                    type: string
                    example: Validation failed (uuid v4 is expected)
                  error:
                    type: string
                    example: Bad request
        404:
          description: User not found
          content:
            application/json:
              schema:
                type: object
                title: example
                properties:
                  statusCode:
                    type: number
                    example: 404
                  message:
                    type: string
                    example: User not found
                  error:
                    type: string
                    example: Not Found
    put:
      tags:
        - Users
      summary: Update a user's password
      description: Updates a user's password by ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              title: example
              properties:
                oldPassword:
                  type: string
                  description: The user's old password
                newPassword:
                  type: string
                  description: The user's new password
              required:
                - oldPassword
                - newPassword
      responses:
        200:
          description: The user has been updated.
          content:
            application/json:
              schema:
                type: object
                title: example
                properties:
                  id:
                    type: string
                    format: uuid
                  login:
                    type: string
                    example: TestUser
                  version:
                    type: integer
                    example: 2
                  createAt:
                    type: integer
                    example: 1655000000
                  updateAt:
                    type: integer
                    example: 1655999999
        400:
          description: Bad request. userId is invalid (not uuid)
          content:
            application/json:
              schema:
                type: object
                title: example
                properties:
                  statusCode:
                    type: number
                    example: 400
                  message:
                    type: string
                    example: Validation failed (uuid v4 is expected)
                  error:
                    type: string
                    example: Bad request
        403:
          description: Forbidden. oldPassword is wrong
          content:
            application/json:
              schema:
                type: object
                title: example
                properties:
                  statusCode:
                    type: number
                    example: 403
                  message:
                    type: string
                    example: Old password is wrong
                  error:
                    type: string
                    example: Forbidden
        404:
          description: User not found
          content:
            application/json:
              schema:
                type: object
                title: example
                properties:
                  statusCode:
                    type: number
                    example: 404
                  message:
                    type: string
                    example: User not found
                  error:
                    type: string
                    example: Not Found
    delete:
      tags:
        - Users
      summary: Delete user
      description: Deletes user by ID.
      responses:
        204:
          description: The user has been deleted
        400:
          description: Bad request. userId is invalid (not uuid)
          content:
            application/json:
              schema:
                type: object
                title: example
                properties:
                  statusCode:
                    type: number
                    example: 400
                  message:
                    type: string
                    example: Validation failed (uuid v4 is expected)
                  error:
                    type: string
                    example: Bad request
        404:
          description: User not found
          content:
            application/json:
              schema:
                type: object
                title: example
                properties:
                  statusCode:
                    type: number
                    example: 404
                  message:
                    type: string
                    example: User not found
                  error:
                    type: string
                    example: Not Found
...
  /album:
    get:
      tags:
        - Albums
      summary: Get albums list
      description: Gets all library albums list
      responses:
        200:
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Album'
        401:
          $ref: '#/components/responses/UnauthorizedError'
    post:
      tags:
        - Albums
      summary: Add new album
      description: Add new album information
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              title: example
              properties:
                name:
                  type: string
                year:
                  type: integer
                artistId:
                  type: string
                  format: uuid
                  nullable: true
              required:
                - name
                - year
      responses:
        201:
          description: Album is created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Album'
        400:
          description: Bad request. body does not contain required fields
          content:
            application/json:
              schema:
                type: object
                title: example
                properties:
                  statusCode:
                    type: number
                    example: 400
                  message:
                    type: string
                    example:
                      - name must be a string
                      - name should not be empty
                      - year must be a number conforming to the specified constraints
                      - year should not be empt
                  error:
                    type: string
                    example: Bad request
  /album/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          example: 0a35dd62-e09f-444b-a628-f4e7c6954f57
    get:
      tags:
        - Albums
      summary: Get single album by id
      description: Gets single album by id
      responses:
        200:
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Album'
        400:
          description: Bad request. albumId is invalid (not uuid)
          content:
            application/json:
              schema:
                type: object
                title: example
                properties:
                  statusCode:
                    type: number
                    example: 400
                  message:
                    type: string
                    example: Validation failed (uuid v4 is expected)
                  error:
                    type: string
                    example: Bad request
        404:
          description: Album was not found.
          content:
            application/json:
              schema:
                type: object
                title: example
                properties:
                  statusCode:
                    type: number
                    example: 404
                  message:
                    type: string
                    example: Album not found
                  error:
                    type: string
                    example: Not Found
    put:
      tags:
        - Albums
      summary: Update album information
      description: Update library album information by UUID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              title: example
              properties:
                name:
                  type: string
                year:
                  type: integer
                artistId:
                  type: string
                  format: uuid
                  nullable: true
              required:
                - name
                - year
      responses:
        200:
          description: The album has been updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Album'
        400:
          description: Bad request. albumId is invalid (not uuid)
          content:
            application/json:
              schema:
                type: object
                title: example
                properties:
                  statusCode:
                    type: number
                    example: 400
                  message:
                    type: string
                    example: Validation failed (uuid v4 is expected)
                  error:
                    type: string
                    example: Bad request
        404:
          description: Album was not found.
          content:
            application/json:
              schema:
                type: object
                title: example
                properties:
                  statusCode:
                    type: number
                    example: 404
                  message:
                    type: string
                    example: Album not found
                  error:
                    type: string
                    example: Not Found
    delete:
      tags:
        - Albums
      summary: Delete album
      description: Delete album from library
      responses:
        204:
          description: Deleted successfully
        400:
          description: Bad request. albumId is invalid (not uuid)
          content:
            application/json:
              schema:
                type: object
                title: example
                properties:
                  statusCode:
                    type: number
                    example: 400
                  message:
                    type: string
                    example: Validation failed (uuid v4 is expected)
                  error:
                    type: string
                    example: Bad request
        404:
          description: Album was not found.
          content:
            application/json:
              schema:
                type: object
                title: example
                properties:
                  statusCode:
                    type: number
                    example: 404
                  message:
                    type: string
                    example: Album not found
                  error:
                    type: string
                    example: Not Found
...
  /favs/album/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          example: 0a35dd62-e09f-444b-a628-f4e7c6954f57
    post:
      tags:
        - Favorites
      summary: Add album to the favorites
      description: Add album to the favorites
      responses:
        201:
          description: Added successfully
        400:
          description: Bad. albumId is invalid (not uuid)
          content:
            application/json:
              schema:
                type: object
                title: example
                properties:
                  statusCode:
                    type: number
                    example: 400
                  message:
                    type: string
                    example: Validation failed (uuid v4 is expected)
                  error:
                    type: string
                    example: Bad request
        422:
          description: Album with id doesn't exist.
          content:
            application/json:
              schema:
                type: object
                title: example
                properties:
                  statusCode:
                    type: number
                    example: 422
                  message:
                    type: string
                    example: Album not found
                  error:
                    type: string
                    example: Unprocessable content
    delete:
      tags:
        - Favorites
      summary: Delete album from favorites
      description: Delete album from favorites
      responses:
        204:
          description: Deleted successfully
        400:
          description: Bad. albumId is invalid (not uuid)
          content:
            application/json:
              schema:
                type: object
                title: example
                properties:
                  statusCode:
                    type: number
                    example: 400
                  message:
                    type: string
                    example: Validation failed (uuid v4 is expected)
                  error:
                    type: string
                    example: Bad request
        404:
          description: Album was not found.
          content:
            application/json:
              schema:
                type: object
                title: example
                properties:
                  statusCode:
                    type: number
                    example: 404
                  message:
                    type: string
                    example: Album not found
                  error:
                    type: string
                    example: Not Found
...